erro issue solvre
