+++ src/app/(pages)/services/page.tsx
@@ -1,11 +1,14 @@
-import React from 'react'
+import ServiceCategory from "@/featrues/services/ServiceCategory";
+import ServicesOverview from "@/featrues/services/ServicesOverview";
+import React from "react";
 
-const Services = () => {
+const Page = () => {
   return (
-    <div>
-      <h1>services</h1>
+    <div className="w-full flex flex-col">
+      <ServicesOverview />
+      <ServiceCategory />
     </div>
-  )
-}
+  );
+};
 
-export default Services
+export default Page;
